DataReaderExtension is moved from BlobStorage to Common
